LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/xhr-in-worker-expected.txt


FAIL Untitled assert_equals: expected "gamma\n" but got "beta\n"
